requirements.txt
langchain[all]
sentence-transformers==5.0.0
flask
pypdf
python-dotenv
pinecone[grpc]
-e . 
